Добавлено списание средств на определенное время, добавлено изменение тарифа(бэкэнд), добавлена история транзакций(бэкэнд)

+++ backend/src/main/java/com/backend/bprobackend/repository/ContractRepos.java
@@ -1,13 +1,13 @@
 package com.backend.bprobackend.repository;
 
 import com.backend.bprobackend.model.Contract;
-import com.backend.bprobackend.model.EnumRole;
-import com.backend.bprobackend.model.Role;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
 import java.util.Optional;
 
-public interface ContractRepos extends JpaRepository<Contract,Long> {
+public interface ContractRepos extends JpaRepository<Contract,Integer> {
     Optional<Contract> findByName(String name);
 
 }
